fix: extract the new strings and reject .tldraw case-insensitively
The messages lived in src/utils/, outside the i18n extraction glob
(src/{tla,pages,components}), so they were never extracted and every
non-English locale would have fallen back to English silently. Move them
under src/tla/utils/ and regenerate the locale files.

File dialogs on macOS and Windows match their filters case-insensitively,
so a .TLDRAW file is selectable and has to be caught too.

The hook now returns the files we can still handle rather than a boolean
no caller read, which removes the duplicated filtering at each call site.
