+++ apps/dotcom/client/src/tla/hooks/useTldrFileDrop.ts
@@ -3,7 +3,7 @@ import { DragEvent, useCallback } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { tlmenus } from 'tldraw'
 import { routes } from '../../routeDefs'
-import { useNotifyTldrawOfflineFiles } from '../../utils/tldrawOfflineFiles'
+import { useRejectTldrawOfflineFiles } from '../utils/tldrawOfflineFiles'
 import { useApp } from './useAppState'
 
 export function useTldrFileDrop() {
@@ -11,7 +11,7 @@ export function useTldrFileDrop() {
 	const navigate = useNavigate()
 
 	const auth = useAuth()
-	const notifyTldrawOfflineFiles = useNotifyTldrawOfflineFiles()
+	const rejectTldrawOfflineFiles = useRejectTldrawOfflineFiles()
 
 	const onDrop = useCallback(
 		async (e: DragEvent) => {
@@ -21,8 +21,7 @@ export function useTldrFileDrop() {
 			}
 
 			if (!e.dataTransfer?.files?.length) return
-			const files = Array.from(e.dataTransfer.files)
-			notifyTldrawOfflineFiles(files)
+			const files = rejectTldrawOfflineFiles(Array.from(e.dataTransfer.files))
 			const tldrawFiles = files.filter((file) => file.name.endsWith('.tldr'))
 			if (!tldrawFiles.length) {
 				return
@@ -31,7 +30,7 @@ export function useTldrFileDrop() {
 				navigate(routes.tlaFile(fileId))
 			})
 		},
-		[app, auth, navigate, notifyTldrawOfflineFiles]
+		[app, auth, navigate, rejectTldrawOfflineFiles]
 	)
 
 	const onDragOver = useCallback((e: DragEvent) => {
